Exercises 4.15.-4.23.

+++ osa4/blog-list/controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if(authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
-
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({})
@@ -20,21 +12,15 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if(!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid '})
-  }
-  const user = await User.findById(decodedToken.id)
-  
-  if (!user) {
-    return response.status(400).json({ error: `user matching id ${body.userId} can't be found` })
+  const loggedInUser = request.user
+  if(!loggedInUser) {
+    return response.status(401).json({ error: 'User not logged in' })
   }
 
-  const blog = new Blog({ ...body, user: user.id }) //Blog gets the user's ID
+  const blog = new Blog({ ...body, user: loggedInUser.id }) //Blog gets the user's ID
   const savedBlog = await blog.save()
-  user.blogs = user.blogs.concat(savedBlog.id) //And user also gets the blog's ID
-  await user.save()
+  loggedInUser.blogs = loggedInUser.blogs.concat(savedBlog.id) //And user also gets the blog's ID
+  await loggedInUser.save()
   response.status(201).json(savedBlog)      
 })
 
@@ -48,12 +34,21 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const blog = await Blog.findByIdAndDelete(request.params.id)
+  const loggedInUser = request.user
+  let blog = await Blog.findById(request.params.id)
+  if(!blog) {
+    return response.status(404).json({ error: "blog not found "})
+  }
+
+  if(loggedInUser.id === blog.user.toString()) {
+    blog = await Blog.findByIdAndDelete(request.params.id)
+  } else {
+    return response.status(401).json({ error: "this user can't delete this blog "})
+  }
+
   if(blog) {
     response.status(204).end()
-  } else {
-    response.status(404).json({ error: "blog not found "})
-  }
+  } 
 })
 
 blogsRouter.put('/:id', async (request, response) => {
